+++ moco-core/src/main/java/com/github/dreamhead/moco/Moco.java
@@ -146,9 +146,20 @@ public final class Moco {
                 mergeMonitor(monitor, monitor2, monitors));
     }
 
+    public static SocketServer socketServer() {
+        return ActualSocketServer.createQuietServer(Optional.<Integer>absent());
+    }
+
     public static SocketServer socketServer(int port) {
         return ActualSocketServer.createQuietServer(of(port));
     }
+
+    public static SocketServer socketServer(final int port, final MocoMonitor monitor) {
+        checkArgument(port > 0, "Port must be greater than zero");
+        return ActualSocketServer.createServerWithMonitor(of(port),
+                checkNotNull(monitor, "Monitor should not be null"));
+    }
+
 
     public static MocoConfig context(final String context) {
         return new MocoContextConfig(checkNotNullOrEmpty(context, "Context should not be null"));