+++ moco-core/src/test/java/com/github/dreamhead/moco/MocoSocketTest.java
@@ -1,18 +1,31 @@
 package com.github.dreamhead.moco;
 
 import com.github.dreamhead.moco.helper.MocoSocketHelper;
+import com.google.common.io.Files;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+import java.io.File;
+import java.nio.charset.Charset;
 
 import static com.github.dreamhead.moco.Moco.by;
+import static com.github.dreamhead.moco.Moco.log;
 import static com.github.dreamhead.moco.Moco.socketServer;
+import static com.github.dreamhead.moco.MocoRequestHit.once;
+import static com.github.dreamhead.moco.MocoRequestHit.requestHit;
 import static com.github.dreamhead.moco.Runner.running;
 import static com.github.dreamhead.moco.helper.RemoteTestUtils.local;
 import static com.github.dreamhead.moco.helper.RemoteTestUtils.port;
+import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
 public class MocoSocketTest {
+    @Rule
+    public TemporaryFolder folder = new TemporaryFolder();
+
     private MocoSocketHelper helper;
     private SocketServer server;
 
@@ -61,6 +74,60 @@ public class MocoSocketTest {
             public void run() throws Exception {
                 helper.connect();
                 assertThat(helper.send(times("a", 1025)), is("long_a"));
+                helper.close();
+            }
+        });
+    }
+
+    @Test
+    public void should_log_request_and_response_into_file() throws Exception {
+        File file = folder.newFile();
+        SocketServer server = socketServer(port(), log(file.getAbsolutePath()));
+        server.request(by("0XCAFE")).response(line("0XBABE"));
+
+        running(server, new Runnable() {
+            @Override
+            public void run() throws Exception {
+                helper.connect();
+                assertThat(helper.send("0XCAFE"), is("0XBABE"));
+                helper.close();
+            }
+        });
+
+        String actual = Files.toString(file, Charset.defaultCharset());
+        assertThat(actual, containsString("0XBABE"));
+        assertThat(actual, containsString("0XCAFE"));
+    }
+
+    @Test
+    public void should_monitor_socket_server_behavior() throws Exception {
+        RequestHit hit = requestHit();
+        SocketServer server = socketServer(port(), hit);
+        server.request(by("0XCAFE")).response(line("0XBABE"));
+
+        running(server, new Runnable() {
+            @Override
+            public void run() throws Exception {
+                helper.connect();
+                assertThat(helper.send("0XCAFE"), is("0XBABE"));
+                helper.close();
+            }
+        });
+
+        hit.verify(by("0XCAFE"), once());
+    }
+
+    @Test
+    public void should_create_socket_server_without_specific_port() throws Exception {
+        final SocketServer server = socketServer();
+        server.request(by("foo")).response(line("bar"));
+
+        running(server, new Runnable() {
+            @Override
+            public void run() throws Exception {
+                helper = new MocoSocketHelper(local(), server.port());
+                helper.connect();
+                assertThat(helper.send("foo"), is("bar"));
                 helper.close();
             }
         });